refactor: move CustomAttachments stories into __stories__ subfolder
