+++ libs/conversation-view/src/components/Attachments/CustomAttachments/__stories__/CustomGrid.stories.tsx
@@ -1,9 +1,9 @@
 import type { Decorator, Meta, StoryObj } from '@storybook/react-vite';
 import { AllCommunityModule, ModuleRegistry } from 'ag-grid-community';
-import { CustomDataGridAttachment } from './CustomGridAttachment';
-import { CustomGridAttachment } from '../../../models/attachments';
-import { ConversationViewStylesProvider } from '../../../context/ConversationViewStylesContext';
-import { OnboardingProvider } from '../../../context/OnboardingContext';
+import { CustomDataGridAttachment } from '../CustomGridAttachment';
+import { CustomGridAttachment } from '../../../../models/attachments';
+import { ConversationViewStylesProvider } from '../../../../context/ConversationViewStylesContext';
+import { OnboardingProvider } from '../../../../context/OnboardingContext';
 
 ModuleRegistry.registerModules([AllCommunityModule]);
